requirements.txt
Django==4.2.10
tensorflow==2.15.0
keras==2.15.0
Pillow>=9.0.0
numpy>=1.23.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
django-rest-framework-simplejwt==5.3.0
bcrypt==4.0.1